audio/filter/delay: Add frame mode
Configured as `delay:24:frames`. In this mode, audio is delayed by the
specified amount of "audio frames". Audio frame is a collection of audio
samples that is read from the audio capture device at once (also called  period size, quantum, etc.)
In the case of decklink embedded audio this corresponds to the video
frame time.