lavc: renamed thread_count to conv_thread_count
This reverts the opposite change in c8dcee51.

Now the conv_thread_count is used solely for pixel format conversion
(was also used as a thread count for VP9 until previous commit).

Also do not deduce conv. thread count from codec thread count (it may
not be related at all - eg. user may want to specify "1" due to some bug
in codec or decoder restriction and then it will be pointless to
restrict the conversion thread count to only one).