loop unrolling clang pragma: use "assume_safety"
Use "assume_safety" for vectorize argument - the conversions for which
we use it are usually pretty data-independent. However clang sometimes
fail to check the safety check (on macOS), see also
'-Rpass-analysis=loop-vectorize'. Perhaps only because the loops are a
bit complicated? Setting all pointers (including auxilliary) as
restricted doesn't seem to help.