GL: fixed irregular sized v210 display
fixed `-d gl` of `-t testcard:size=47x32:codec=v210` or `..size=48x32`

The problem was perhaps with truncation of a number slightly below
nearest integer. Result was a "broken" bar of a testcard (in midst it
was shifted by one horizontal pixel).

Quite a huge eps is needed in macOS - .49999 cut-off was not sufficient
for `size=$((48*53))x$((32*53))`.