configure: Check for vulkan headers
The vulkan.pc pkgconfig file is for the vulkan loader library and
doesn't guarantee the presence of the vulkan headers. Indeed, some
distributions (e.g. Arch) package those in separate packages
(vulkan-icd-loader and vulkan-headers).

It would be perhaps more correct to check for vulkan.hpp, but that
header is so big that it would slow the configuration by ~10 seconds.