lavc: removed AVCodec from pixfmt offer
As it is written now, passing the AVCodec parameter was unneeded since
we are returning with appropriate function rather the list of all AV
pixels format that are possible to use without respect to actual codec.

Actually it was only used to check `vaapi` in codec name - this is not
needed, we can add vaapi to the list regardless to it - if the codec
doesn't support this, it just skips it.