video dec: change chosen disp pixfmt from compress
Use compare_pixfmt for codec sorting resulting in different selection,
eg:

    uv -t testcard:codec=R10k -c libavcodec:encoder=libx265:yuv -d dummy:codec=decklink

Used to end with v210 (compressed properties was 10-bit YUV 4:4:4;
internal codec detected as Y416). Now it ends with R12L - note that this
is still not optimal as it should be rather R10k. For GL it even ends up
with RG48.

The previous behavior can be triggered by `--conv-policy cds`.

refer to GH-295