lavc audio: suppress a warning
suppress missing-field-initializers warning when assigning
AVChannelLayout to AVCodecContext::ch_layout. This is C++ specific, the
construction is entirely fine in C (empty-initializes remaining
members).

Rewritten to C - it seems to be a bit invasive, because the rewrite is
quite huge. On the other hand, it cleans the code a bit and also
removes some inefficiencies that there have been (now not possible
because the absence of RAII).