recreate the rendering pipeline if size changes
If shader conversion is used, the rendering pipeline needs to be
recreated because the converted frames are initialized with the actual
size. We can perhaps recreate it always if the properties change,
although strictly taken it is not necessarily needed if there is no
conversion and only size changes.

refer to GH-303