Win CI scripts: create /usr/local/* if not
Ensure that /usr/local/{bin,include,lib} are created in scrips that copy
to that.

This is not problem for the CI because `.github/scripts/Windows/prepare_msys.sh`
already creates them but the scripts are standalone installers and it
can be run independently by a developer to install the particular dependency.