hd-rum-translator: fixed messages not delivered
The state_root::new_message is now handled in host.cpp, which processed
all messages, even those intended for reflector.

As a solution, the new_message callback of state_root processes only
messages for it and puts remaining messages back to the pool with the
new function module_store_message() that doesn't call the new_message
callback.