sdl_mixer: prefer system sound font
check if Mix_GetSoundFonts() is a usable font and if so, use

This has precedence over bundled songs. But most importantly, it will
allow running also from just compiled source, where there is no sound
font relative to `uv` executable but there could be a sound font in
system paths.