CoreAudio: print error FourCC if available
In addition to error code, the OSStatus usually represent a FourCC that
can be presented to user.

Some generic error may be interpreted by:

    NSError *error = [NSError errorWithDomain:NSOSStatusErrorDomain code:err userInfo:nil];
    const char *err_str = [[error localizedDescription] UTF8String];

But it looks like it almost always returns 'The operation couldn’t be
completed.' for unknown errors. Also there doesn't seem to be any domain
specific for CoreAudio.