vidcap aggregate: pass parent to slaves
See also d1d4843e (similar change for switcher). This patch is slightly
different because the slaves are connected to aggregate parent (it
doesn't initialize own module) whereas switcher has the state and acts
as a parent to its slaves. Shouldn't make much difference if path to the
module isn't important.

This fixes vidcap _file_ requiring root to register keybindings.