vidcap file: increase audio ring buf size to 60 sec
It needs to be more than GOP size if audio is +/- aligned with video,
eg. The Elephants Dream has GOP 150 frames (= 6.25 sec at 24 FPS)
meaning that the audio must wait up to more than 7 seconds (in the file,
audio packets are up to 1 sec delayed to the respective video packets).