vidcap file worker: do house-keeping first
do the house-keeping first and remove unused CV

The queue is already written over the max when there are video frames
waiting for audio (and there may be even something like 150, like in
ED[1]) so it doesn't matter if we write one more frame or not.

This simplifies the stuff a bit.

+ it allows to check the messages always (not only before frames) -
  again this is useful when ED flushes some 150 frames, because before
  this change, it was processed just after all those frames were sent
+ print 'paused' notice when the stream stops (without loop) - the
  stream can be rewinded back, but it is in paused state, so unpause
  must be issued at the same time

[1] The Elephants Dream