use lrand48 instead of lbl_random
Custom lbl_random implementation is not properly seeded, when not in
Windows, where it is used instead of [ld]rand48. rand48 functions on the
other hand are seeded (in host.cpp and rtp.c) so it can be used instead
(it perhaps doesn't make much sense to use both implementation in
UltraGrid except Windows, where custom implementation is used solely).

lbl_[s]random remains used in crypto_des (which is, however, not used).