Makefile: rewritten macOS bundling
Always create clean bundles (both with and without GUI).

The former approach was quite fragile and althoug it offered some
advantage that it wasn't needed to rebundle the libs all the time, it
perhaps doesn't overcome the disadvantages, eg. that it was actually
broken until now - first run was successfull but when the macos-wrapper
was deployed second time, it created broken bundle (the script doesn't
handle this case).

Some kind of caching of already processed steps may be readded in future
but definitely made in a more clean way.

+ removed bundle-nolib - doesn't seem to be needed; can be readded if
  needed, anyways
+ removed '; \' at line before 'cp -R $(GUI_BUNDLE_DEP)/* $(GUI_BUNDLE)'
  Not sure how it affects the behavior but that made the cp command
  part of the previous command, which is certainly not intended.