DeckLink disp.: reconfigure audio w. video (schd)
In scheduled mode, we need to restart stream after configuring audio. The
previous implemenation was rather error-prone because if the audio was
configured after video, it didn't work. Now it simplifies the workflow.

This also fixes a potential crash when DisableVideoOutput in reconfigure
was called after deleting the delegate structure.