Revert "lavd video: don't print false positive errs"
This reverts commit d4c07ee44faa157054870f0c34f3dbad4daf8cdf.

The fix was not correct, because it just hided the root of the problem -
the av_parser_parse2 needs final call with buf_size == 0 to flush last
frame. Otherwise it will remain in the queue, effectivelly adding a
delay by one frame, because the particular frame will get flushed when
processing next frame.

See also:
<https://github.com/FFmpeg/FFmpeg/blob/master/doc/examples/decode_video.c>