GIT CI ARM: do not build own FFmpeg
Own FFmpeg was build only to enable periodic insertion of SPS/PPS NALU
for OMX. This should be now, however, possible to do even with the lavc
`header_inserter` option (which is enabled by default for OMX). Also
the OMX API seems unavailable in more recent RPi OS in favor of other
APIs.