GL VDPAU: re-added GL_NV_vdpau_interop check
Re-added check to GL_NV_vdpau_interop removed by mistake in commit
a0c094f0 (21st Apr 2022).

In the previous solution error was written but retval was not checked,
causing a crash. Now only warning is issued and HW_VDPAU is not
offered. In state_vdpau::init() is used assertion.