CI Win: build own Spout
closes GH-352

It looks like the pre-build library BUILD/Binaries/x64/SpoutLibrary.dll
is no longer compatible with the headers in Git.

Pre-built library version is 2.007.010 while the latest version on GitHub
is Spout 2.007.012. When building Spout from sources, the referenced
issue no longer occurs. Also the libraries are not expected to be included
in the Git in future (leadedge/Spout2#103).