CI Win Cineform: do not copy .pc file
After previous commit, UG configure finds the .pc file because
the file was copied incorrectly before - it is copied to
/usr/local/lib/pkgconfig, which was perhaps not a directory, creating
file _named_ pkgconfig instead the file inside that dir. Interestingly
this didn't look like it influenced other libraries, that might also
want to copy its files to something, that was then file.

But the contents is not usable without adjustnments:
1. prefix is C:/Program Files/CineFormSDK
2. lib is set to -lCFHDCodecShared in the PC file, but generated
library has name just CFHDCodec (something similar as already solved
by 0001-CMakeList.txt-remove-output-lib-name-force-UNIX.patch)

While the pkg-config file was not used previously, it is perhaps easiest
to leave it as it used to be instead of trying to fix the .pc file.