cmpto_j2k enc.: use pinned tmp buffers
Frame data is copied, anyways, so if we have CUDA support in UG (not
only transitively through the codec), use CUDA host buffers.

Doesn't seem to have an impact to the throughput with GK110-400-A1
with default pool_size=4 but with pool_size=3, the throughput
increases by more thant 20%, now only 10% slower than p_s=4 (4K FPS 42,
and 38 with ps=3; before this commit 31.5).