lavd: Fix SEGV on unknown sw_pix_fmt (libdav1d)
avcodec only guarantees that sw_pix_fmt is set before calling the
get_format callback. However for libdav1d the callback is for some
reason never called and sw_pix_fmt is not set.

So just use the pix_fmt if sw_pix_fmt is unset. Not sure if this is the
best fix, but it works in this case and doesn't break anything.