vcap/dshow: accept any codec for pos. arg
we can accept all codecs, not only BGR or YUYV/YUY2

Note that the codec is used as both streaming and grabber codecs,
the converting logic doesn't seem to have been working in the past,
although Pavel Troubil's original code had RGB fixed as the codec.

Commented out the block setting the mediaformat - now the matching format
is already set, unsure about semantic of this block.