vpp/delay: limit nr of frames
added a limit to number of frames/seconds that can be cached

The user may enter some huge value by mistake and since FHD UYVY frame
takes 4M, it can exhaust the memory quite quickly. So the limit is set
to 2000 frames or 60 sec, which is in the case of 30p video up to 8 GB
of memory.

\+ print the user selection to console