vcap/v4l2: increase default buffer cnt to 3
The allocated buffer size doesn't mean increasing the latency in case
that the device manages to process the frames. On the other hand, in
the opposite case, it fixes throughput. Actually, in this case it will
be possible to use even more than those 3 buffers, but just keep it
conservative for now.

The actual card for which this setting helps is a cheap one with
MacroSilicon MS2109 chip (USB ID 534d:2109).