common_preinit: use non-permuting getopt
Instead of tweaking the getopt manually, use non-permuting version of
getopt (POSIX-compiliant, indicated by leading '+' in optstring).

This is slightly less error-prone and more readable. Althoug still a
bit tweaking is needed - skipping a non-opt (which can be in the middle
of argv if user cmd uses this syntax). Also options for opts that are
not recognized by this getopt (any aother than -O/-V) will appear as
a non-opt.