configure.ac: added all to --enable-object
The options cannot be repeated for configure, like
`--enable-object=a --enable-object=b` to enable both a and b
(`--enable-object=a,b` is needed). This may be sometimes impractical,
when the command-line args are appended incrementally.

As a workaround, at least `--enable-output=all` is added. As it is not
a solution, it still (with `--disable-all`) removes most unneeded code
(with dependencies), while compiling all the non-essential objs without
deps.