com_initialize: warn if not false
The guard value should be used only for one pair of
com_initialize/uniniialize call, so ensure it by checking if the value
is initialized to false.

It should be possible to use an int instead but there is currently
no need.

Set the occurences to false explicitly (in the case of vcap/deck change
from true to false).