Revert "fixed CID 424197" + fix correctly
Fix the commit 31ea07470 (2023-11-23) that fixed (incorrectly) the
CID 424197.

If  should_exit was set before waiting in pthread_cond_timedwait, the
infinite loop is never left. Moreover, it seems that the loop was left
only on should_exit timeout, not when notified by condition variable.

This is mostly to make Coverity happy - the spurious wakeup here does
not matter in any way. So reverting the original "fix" would be enough.