vcap/rtsp: do the SDP parsing more reliably
The original mean was a bit cumbersome so rewrite it entirely in a more clean way.

It is now capable of playing the video produced by this UG command:
```
uv -t testcard -x rtsp:port=8654 -c lavc:enc=libx264:safe -V -A mp3 -s testcard
```

Note that the audio is still not fully implemented so that only video will
be presented. Neither this was possible before this change (audio control
was used instead of video due to buggy parsing).