vcap/rtsp: improved error handling
On RTSP error, return from the header callback CURL_WRITEFUNC_ERROR
to signalize cURL library that we have a problem. The particular cURL
function will then fail accordingly and we don't need to handle with
an external variable.

As a conseqence, all errors are now handled as a failure, not only
when rtsp_error_occurred was checked.