vdisp: fixed restrict_returned_codecs
Rewritten restrict_returned_codecs to respect eventual pixfmt change by
the postprocessor (as with change_pixfmt).

Originally the set intersection of codecs accepted for vo_postprocess
and display was used, which is unsuitable when the vo_postprocess changes
the format. Now the postprocessor is queried for the output format instead,
which is then compared with display list of supported codecs.