color.h: give up clamp for limited value
The [document] referenced in the header is far from being strict in this
respect. The values that were clamed to was Nominal Video Range. At
the same time, the Preferred Min./Max. is significantly lower/higher
(16-235 vs 5-246 for 8 bit). This value can be understood as a "soft"
limit while the Total Video Signal Range (1-254) as a hard limit.

Some decoders (FFmpeg HEVC) overshot the nominal values, anyways.

[document]:
https://tech.ebu.ch/docs/r/r103.pdf