bmd_status: use module prefix
to be able to identify capture vs playback

This may or may not be important - if cap/disp is on single device,
the status is global so it doesn't matter there. But those can be 2
separate DeckLinks and then there will be hard to differntiate.

Also, both devices will get the same notifications so this makes at least
clear that the duplicate has some reason. (Maybe it will be useful to
figure out how to print just once?)

Note that BMD_CHECK still uses the blackmagic_common.cpp MOD_NMAE
("[DeckLink] ").