net_udp: fix mcast TTL not set also for IPv6
This fixes the commit 0b557409 (2021-03-29) that added this for both
IPv4 and IPv6 but for IPv6 the condition was incorrectly aligned to
setsockopt fail, not ttl==-1.

Dropped IPv4/IPv6 specification - since we may/will use v4-mapped IPv6
sockets, the word "IPv6" might be confusing if use is using IPv4 mcast
addr.