configure.ac: add option not to use MoltenVK [mac]
On mac, we can link either to MoltenVK or the vulkan-loader. The MoltenVK
variant is default, because the bundler then bundles the library as the
compat layer. If linked to vulkan-loader, Vulkan will likely not work if
the MoltenVK is not installed on the host computer.

But it is useful for debugging (the behavior is slightly different if
loaded via vulkan-loader even if the backend is MoltenVK - the KHR
portability bit/enumeration stuff).