vdec/lavc: print frame flags in debug
Frame type (I/P/B) is already printed in debug so add also frame flags
which may be useful (eg. for the recent x265 intra-refresh decode
fix  7bed4ee5 - flag indicates corrupted frame all the time, which
may or may not be an incorrect behavior from hevc decoder; but doesn't
matter for us, anyways).

Also use MSG(), which also has a advantage that if log_level not met,
it is not evaluated.