vcap/deltacast: request channel idx + 1 RX channels
For bidirectional cards (eg. deltacast hd elp-d 80), the channels can be
configured in a layout that lower indices are assigned to RX and higher
to TX (so 0-8 lower-indexed pins are RX and the rest is TX).

So if using RX channel=2 (zero-indexed), at least 3 RX channels need
to be taken in order for the 3rd channel to be RX.